console: Change location of handlers
