+++ pkg/webui/console/containers/invite-user-form/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,18 +14,27 @@
 
 import React, { useState } from 'react'
 import { defineMessages } from 'react-intl'
+import { connect } from 'react-redux'
+import { push } from 'connected-react-router'
 
 import Form from '@ttn-lw/components/form'
 import Input from '@ttn-lw/components/input'
 import SubmitBar from '@ttn-lw/components/submit-bar'
 import SubmitButton from '@ttn-lw/components/submit-button'
 
+import Message from '@ttn-lw/lib/components/message'
+
 import Yup from '@ttn-lw/lib/yup'
 import PropTypes from '@ttn-lw/lib/prop-types'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
+import attachPromise from '@ttn-lw/lib/store/actions/attach-promise'
+
+import { sendInvite } from '@console/store/actions/users'
 
 const m = defineMessages({
   emailPlaceholder: 'mail@example.com',
+  invitationsDescription:
+    'You can invite users to this network by providing an email address. The person will then get an email with instructions on how to join your network.',
 })
 
 const validationSchema = Yup.object().shape({
@@ -33,12 +42,16 @@ const validationSchema = Yup.object().shape({
 })
 
 const InviteForm = props => {
-  const { onSubmit, onSubmitSuccess } = props
+  const { sendInvite, navigateToList } = props
+
+  const onSubmit = React.useCallback(async email => sendInvite(email), [sendInvite])
+
+  const onSubmitSuccess = React.useCallback(() => navigateToList(), [navigateToList])
 
   const [error, setError] = useState()
   const handleSubmit = React.useCallback(
     async (values, { resetForm, setSubmitting }) => {
-      setError('')
+      setError(undefined)
       try {
         const result = await onSubmit(values)
         resetForm({ values })
@@ -56,33 +69,36 @@ const InviteForm = props => {
   }
 
   return (
-    <Form
-      error={error}
-      onSubmit={handleSubmit}
-      initialValues={initialValues}
-      validationSchema={validationSchema}
-    >
-      <Form.Field
-        title={sharedMessages.emailAddress}
-        component={Input}
-        name="email"
-        placeholder={m.emailPlaceholder}
-        required
-      />
-      <SubmitBar>
-        <Form.Submit message={sharedMessages.invite} component={SubmitButton} />
-      </SubmitBar>
-    </Form>
+    <>
+      <Message content={m.invitationsDescription} component="p" />
+      <hr className="mb-ls-m" />
+      <Form
+        error={error}
+        onSubmit={handleSubmit}
+        initialValues={initialValues}
+        validationSchema={validationSchema}
+      >
+        <Form.Field
+          title={sharedMessages.emailAddress}
+          component={Input}
+          name="email"
+          placeholder={m.emailPlaceholder}
+          required
+        />
+        <SubmitBar>
+          <Form.Submit message={sharedMessages.invite} component={SubmitButton} />
+        </SubmitBar>
+      </Form>
+    </>
   )
 }
 
 InviteForm.propTypes = {
-  onSubmit: PropTypes.func.isRequired,
-  onSubmitSuccess: PropTypes.func,
+  navigateToList: PropTypes.func.isRequired,
+  sendInvite: PropTypes.func.isRequired,
 }
 
-InviteForm.defaultProps = {
-  onSubmitSuccess: () => null,
-}
-
-export default InviteForm
+export default connect(null, {
+  sendInvite: email => attachPromise(sendInvite(email)),
+  navigateToList: () => push(`/admin/user-management`),
+})(InviteForm)
