+++ pkg/webui/lib/shared-messages.js
@@ -367,6 +367,7 @@ export default defineMessages({
   secondsAbbreviated: 'sec',
   secret: 'Secret',
   secure: 'Secure',
+  sendInvitation: 'Send invitation',
   sessions: 'Sessions',
   serviceData: 'Service data',
   settings: 'Settings',
@@ -418,6 +419,7 @@ export default defineMessages({
   userDelete: 'Delete user',
   userEdit: 'Edit user',
   userId: 'User ID',
+  userInvitations: 'User invitations',
   userManagement: 'User management',
   username: 'Username',
   users: 'Users',
